Normalize to current bps

+++ src/interfaces/IAuction.sol
@@ -51,7 +51,7 @@ interface IAuction is ITickStorage, IAuctionStepStorage {
     /// @param amount The amount of the bid
     /// @param owner The owner of the bid
     /// @param prevHintId The id of the previous tick hint
-    function submitBid(uint128 maxPrice, bool exactIn, uint128 amount, address owner, uint128 prevHintId)
+    function submitBid(uint128 maxPrice, bool exactIn, uint256 amount, address owner, uint128 prevHintId)
         external
         payable;
 }
